DC007
atualização da estrutura
