+++ mobile/lib/screens/profile_screen.dart
@@ -1,0 +1,101 @@
+import 'package:flutter/material.dart';
+import '../widgets/logo_widget.dart';
+
+class ProfileScreen extends StatelessWidget {
+  const ProfileScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: const Color(0xFF26A69A),
+        elevation: 0,
+        leading: IconButton(
+          icon: const Icon(Icons.arrow_back, color: Colors.white),
+          onPressed: () {
+            Navigator.pop(context);
+          },
+        ),
+        actions: [
+          TextButton(
+            onPressed: () {
+              Navigator.of(context).pushReplacementNamed('/home');
+            },
+            child: const Text('Menu', style: TextStyle(color: Colors.white)),
+          ),
+        ],
+      ),
+      body: Column(
+        children: [
+          const LogoWidget(),
+          const SizedBox(height: 20),
+          const CircleAvatar(
+            radius: 50,
+            backgroundImage: NetworkImage('https://via.placeholder.com/100'),
+          ),
+          const SizedBox(height: 24),
+          _buildProfileItem('Nome', 'Fulano da Silva'),
+          const SizedBox(height: 12),
+          _buildProfileItem('E-mail', 'Fulanodasilva@gmail.com'),
+          const SizedBox(height: 12),
+          _buildProfileItem('Permissões', 'Admin'),
+          const Spacer(),
+          const Divider(color: Colors.white30),
+          Container(
+            width: double.infinity,
+            padding: const EdgeInsets.symmetric(
+              horizontal: 16.0,
+              vertical: 8.0,
+            ),
+            color: Colors.white,
+            child: const Row(
+              children: [
+                Text(
+                  'Perfil',
+                  style: TextStyle(
+                    fontSize: 18,
+                    fontWeight: FontWeight.bold,
+                    color: Colors.black87,
+                  ),
+                ),
+                Spacer(),
+                Icon(Icons.more_vert),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildProfileItem(String label, String value) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 16.0),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            label,
+            style: const TextStyle(color: Colors.white70, fontSize: 14),
+          ),
+          const SizedBox(height: 4),
+          Container(
+            width: double.infinity,
+            padding: const EdgeInsets.symmetric(
+              horizontal: 16.0,
+              vertical: 12.0,
+            ),
+            decoration: BoxDecoration(
+              color: Colors.white,
+              borderRadius: BorderRadius.circular(8.0),
+            ),
+            child: Text(
+              value,
+              style: const TextStyle(color: Colors.black87, fontSize: 16),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
