+++ mobile/lib/widgets/menu_card.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+class MenuCard extends StatelessWidget {
+  final String title;
+  final VoidCallback onTap;
+
+  const MenuCard({super.key, required this.title, required this.onTap});
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: onTap,
+      child: Container(
+        padding: const EdgeInsets.symmetric(horizontal: 16.0),
+        decoration: BoxDecoration(
+          color: Colors.white,
+          borderRadius: BorderRadius.circular(8.0),
+        ),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Text(
+              title,
+              style: const TextStyle(
+                fontSize: 18,
+                fontWeight: FontWeight.w500,
+                color: Colors.black87,
+              ),
+            ),
+            const Icon(Icons.more_vert, color: Colors.black54),
+          ],
+        ),
+      ),
+    );
+  }
+}
